Covid-Messiah/requirements.txt
absl-py==0.9.0
alembic==1.5.8
altgraph==0.17
appdirs==1.4.3
astroid==2.4.2
astropy==4.0.1.post1
astunparse==1.6.3
atari-py==0.2.6
attrs==19.3.0
backcall==0.1.0
bcrypt==3.2.0
beautifulsoup4==4.9.1
bleach==3.1.3
branca==0.4.0
cachetools==4.1.1
certifi==2019.11.28
cffi==1.14.3
chardet==3.0.4
click==7.1.1
cliff==3.7.0
cloudpickle==1.3.0
cmaes==0.8.2
cmd2==1.5.0
colorama==0.4.3
colorlog==4.8.0
colorlover==0.3.0
config==0.5.0
confuse==1.3.0
cufflinks==0.17.3
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
dialogflow==1.1.0
distlib==0.3.0
entrypoints==0.3
filelock==3.0.12
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-Migrate==2.7.0
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
folium==0.10.1
future==0.18.2
gast==0.3.3
google-api-core==1.26.3
google-auth==1.29.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
googleapis-common-protos==1.53.0
grpcio==1.31.0
gunicorn==20.0.4
gym==0.17.2
h5py==2.10.0
html5lib==1.1
htmlmin==0.1.12
idna==2.9
ImageHash==4.1.0
ipykernel==5.2.0
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==5.6.4
itsdangerous==1.1.0
jedi==0.16.0
Jinja2==2.11.1
joblib==0.16.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.0
jupyter-console==6.1.0
jupyter-core==4.6.3
jupyterthemes==0.20.0
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
lesscpy==0.14.0
llvmlite==0.34.0
lxml==4.5.2
Mako==1.1.4
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.3.0
mccabe==0.6.1
missingno==0.4.2
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.4
networkx==2.4
nltk==3.5
notebook==6.0.3
numba==0.51.0
numpy==1.18.2
oauthlib==3.1.0
opencv-python==4.2.0.32
opt-einsum==3.3.0
optuna==2.7.0
packaging==20.9
pandas==1.0.3
pandas-datareader==0.9.0
pandas-profiling==2.8.0
pandocfilters==1.4.2
parso==0.6.2
pbr==5.5.1
pefile==2019.4.18
phik==0.10.0
pickleshare==0.7.5
Pillow==7.2.0
plotly==4.9.0
ply==3.11
praw==6.5.1
prawcore==1.0.1
prettytable==2.1.0
prometheus-client==0.7.1
prompt-toolkit==3.0.4
protobuf==3.13.0
psycopg2==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyglet==1.5.0
Pygments==2.6.1
PyInstaller==3.6
pylint==2.6.0
pyparsing==2.4.7
pyperclip==1.8.2
pyreadline3==3.3
pyrsistent==0.16.0
PySocks==1.7.1
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2019.3
PyWavelets==1.1.1
pywin32==227
pywin32-ctypes==0.2.0
pywinpty==0.5.7
PyYAML==5.3.1
pyzmq==19.0.0
qtconsole==4.7.1
QtPy==1.9.0
regex==2020.7.14
requests==2.23.0
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.6
scikit-learn==0.23.2
scipy==1.4.1
seaborn==0.10.1
selenium==3.141.0
Send2Trash==1.5.0
six==1.14.0
soupsieve==2.0.1
SQLAlchemy==1.3.16
stevedore==3.3.0
tangled-up-in-unicode==0.0.6
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.0
tensorflow-estimator==2.3.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
threadpoolctl==2.1.0
toml==0.10.1
torch==1.6.0
torchvision==0.7.0
tornado==6.0.4
tqdm==4.48.2
traitlets==4.3.3
tweepy==3.10.0
update-checker==0.16
urllib3==1.25.8
virtualenv==20.0.18
visions==0.4.4
wcwidth==0.1.9
webencodings==0.5.1
websocket-client==0.57.0
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wordcloud==1.8.0
wrapt==1.12.1
wtf==0.1
WTForms==2.2.1
xgboost==1.3.3
